refactor(engine): split sentinel operations.ts into ops/ modules
Splits the 811-line operation registry into 6 focused modules:
- ops/registry.ts (OperationFn type, registry, getOperation, executeOperation)
- ops/d1-ops.ts (D1 read/write operation generators)
- ops/kv-ops.ts (KV get/put operation generators)
- ops/r2-ops.ts (R2 upload/download operation generators)
- ops/domain-ops.ts (auth, post CRUD, media operation generators)
- ops/cleanup.ts (cleanupSentinelData)

Replaces side-effect registry population with explicit registerOperations()
calls in the index barrel. Original operations.ts becomes a thin re-export.
